+++ app/Http/Controllers/Api/FilesControlller.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Models\Admin;
 use App\Models\File;
 use App\Models\Subject;
 use Illuminate\Http\Request;
@@ -64,28 +65,41 @@ class FilesControlller extends Controller
 
     public function deletePdf($id)
     {
+        $admin_id = auth()->user()->id ;
+        $admin = Admin::find($admin_id);
+        // $find = Admin::where('role','super_admin');
+        // ->Subject::where('created_by',$admin_id)->first();
+        $sup = File::find($id);
 
-        // Find the PDF object
-        $pdf = File::find($id);
+        if($admin->role == 'super_admin' || $sup->created_by == $admin_id ){
 
-        if (!$pdf) {
+            // Find the PDF object
+            $pdf = File::find($id);
+
+            if (!$pdf) {
+                return response()->json([
+                    'status' => false,
+                    'message' => 'الملف غير موجود'
+                ], 404);
+            }
+
+            // Delete the PDF file from storage
+            if (file_exists(public_path('files/' . $pdf->file))) {
+                unlink(public_path('files/' . $pdf->file));
+            }
+
+            // Delete the PDF record from the database
+            $pdf->delete();
+
             return response()->json([
-                'status' => false,
-                'message' => 'الملف غير موجود'
-            ], 404);
+                'status' => true,
+                'message' => 'تم حذف الملف بنجاح'
+            ]);
+        }else{
+            return response()->json([
+                'status' => true,
+                'message' => 'ليس لديك الصلاحية للقيام بذلك '
+            ],402);
         }
-
-        // Delete the PDF file from storage
-        if (file_exists(public_path('files/' . $pdf->file))) {
-            unlink(public_path('files/' . $pdf->file));
-        }
-
-        // Delete the PDF record from the database
-        $pdf->delete();
-
-        return response()->json([
-            'status' => true,
-            'message' => 'تم حذف الملف بنجاح'
-        ]);
     }
 }
