+++ app/Http/Controllers/Api/OptionsControlller.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
+use App\Models\Admin;
 use App\Models\Option;
 use App\Models\Question;
 use Illuminate\Http\Request;
@@ -10,6 +11,7 @@ use Illuminate\Support\Facades\Validator;
 
 class OptionsControlller extends Controller
 {
+    
     public function setOption(Request $request , $id )
     {
         //Validations
@@ -70,20 +72,33 @@ class OptionsControlller extends Controller
     }
 
     public function deleteOption($option_id){
-        $option_deleted = Option::find($option_id);
+        $admin_id = auth()->user()->id ;
+        $admin = Admin::find($admin_id);
+        // $find = Admin::where('role','super_admin');
+        // ->Subject::where('created_by',$admin_id)->first();
+        $sup = Option::find($option_id);
 
-        if($option_deleted){
-            $option_deleted->delete();
+        if($admin->role == 'super_admin' || $sup->created_by == $admin_id ){
+            $option_deleted = Option::find($option_id);
 
-            return response()->json([
-                'status'=>true,
-                'message'=>'تم حذف الخيار بنجاح '
-            ]);
+            if($option_deleted){
+                $option_deleted->delete();
+
+                return response()->json([
+                    'status'=>true,
+                    'message'=>'تم حذف الخيار بنجاح '
+                ]);
+            }else{
+                return response()->json([
+                    'status'=>false,
+                    'message'=>"هذا الخيار غير موجود "
+                ],404);
+            }
         }else{
             return response()->json([
-                'status'=>false,
-                'message'=>"هذا الخيار غير موجود "
-            ],404);
+                'status' => true,
+                'message' => 'ليس لديك الصلاحية للقيام بذلك '
+            ],402);
         }
 
     }
